New fitting code
