+++ smpy/__init__.py
@@ -1,4 +1,4 @@
 from smpy import *
 from .version import __version__
 
-__all__ = [ 'ssp', 'dust', 'sfh']
+__all__ = [ 'ssp', 'dust', 'sfh', 'misc']
